+++ examples/paginate/go.mod
@@ -7,10 +7,10 @@ replace github.com/stackitcloud/stackit-sdk-go/services/automation => ../../serv
 
 replace github.com/stackitcloud/stackit-sdk-go/services/albwaf => ../../services/albwaf
 
-replace github.com/stackitcloud/stackit-sdk-go/core => ../../core
+replace github.com/stackitcloud/stackit-sdk-go/experimental => ../../experimental
 
 require (
-	github.com/stackitcloud/stackit-sdk-go/core v0.26.0
+	github.com/stackitcloud/stackit-sdk-go/experimental v0.1.0
 	github.com/stackitcloud/stackit-sdk-go/services/albwaf v0.13.2
 	github.com/stackitcloud/stackit-sdk-go/services/automation v0.1.0
 )
@@ -18,4 +18,5 @@ require (
 require (
 	github.com/golang-jwt/jwt/v5 v5.3.1 // indirect
 	github.com/google/uuid v1.6.0 // indirect
+	github.com/stackitcloud/stackit-sdk-go/core v0.26.0 // indirect
 )
