+++ experimental/NOTICE.txt
@@ -1,0 +1,2 @@
+STACKIT Experimental SDK for Go
+Copyright 2024 - 2026 Schwarz IT KG
