+++ experimental/go.mod
@@ -1,0 +1,5 @@
+module github.com/stackitcloud/stackit-sdk-go/experimental
+
+go 1.25
+
+require github.com/google/go-cmp v0.7.0
